Shift to cache as primary approach
